Various enhancements and fixes: * extended circuit status (all/day/night/off) * fixed day/night adjustment handling * heating level computing algorithm fixed * running with args * manager security enhancements * tests * package building

+++ HomeHeatManager/src/main/java/com/sasieczno/homeheat/manager/security/RefreshTokenAuthenticationProvider.java
@@ -1,13 +1,17 @@
 package com.sasieczno.homeheat.manager.security;
 
 import com.sasieczno.homeheat.manager.repository.TokenRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.stereotype.Component;
 
+import java.util.concurrent.atomic.AtomicLong;
+
 @Component
+@Slf4j
 public class RefreshTokenAuthenticationProvider implements AuthenticationProvider {
 
     @Autowired
@@ -19,26 +23,33 @@ public class RefreshTokenAuthenticationProvider implements AuthenticationProvide
     @Override
     public Authentication authenticate(Authentication authentication) throws AuthenticationException {
         if (authentication instanceof RefreshTokenAuthentication) {
+            log.debug("RefreshTokenAuthentication.authenticate: {}", authentication);
             AuthData authData = tokenRepository.getAuthDataByRefreshToken((String)authentication.getPrincipal());
+            log.debug("RefreshTokenAuthentication.authenticate authData: {}", authData);
             if (authData == null)
                 throw new RefreshTokenAuthenticationException("Invalid or expired refresh token");
-            if (((RefreshTokenAuthentication) authentication).isLogout()) {
-                tokenRepository.removeAuthData(authData);
-            } else {
-                authData.setToken(tokenService.generateAccessToken(authData.getUsername()));
-                tokenRepository.saveAuthData(authData);
-                ((RefreshTokenAuthentication) authentication).setCredentials(authData.getToken());
+            synchronized (authData) {
+                if (!authData.isValid())
+                    throw new RefreshTokenAuthenticationException("Invalid or expired refresh token");
+                if (((RefreshTokenAuthentication) authentication).isLogout()) {
+                    tokenRepository.removeAuthData(authData);
+                    authData.setValid(false);
+                } else {
+                    tokenRepository.removeAuthData(authData);
+                    authData.setRefreshToken(tokenService.generateRefreshToken(new AtomicLong(authData.getExpiryDate())));
+                    authData.setToken(tokenService.generateAccessToken(authData.getUsername()));
+                    tokenRepository.saveAuthData(authData);
+                    ((RefreshTokenAuthentication) authentication).setCredentials(authData);
+                }
+                authentication.setAuthenticated(true);
+                return authentication;
             }
-            authentication.setAuthenticated(true);
-            return authentication;
         }
         return null;
     }
 
     @Override
     public boolean supports(Class<?> aClass) {
-        if (RefreshTokenAuthentication.class.isAssignableFrom(aClass))
-            return true;
-        return false;
+        return RefreshTokenAuthentication.class.isAssignableFrom(aClass);
     }
 }
